MAINT use size_t in Liblinear
Was: d2f18015169e9981f82fb520af9532e1dbccecf3

replace long int with size_t in linear.cpp and train.c

Conflicts:
	sklearn/svm/src/liblinear/linear.cpp
	train.c

+++ sklearn/svm/src/liblinear/linear.cpp
@@ -2076,8 +2076,8 @@ static void transpose(const problem *prob, feature_node **x_space_ret, problem *
 	int i;
 	int l = prob->l;
 	int n = prob->n;
-	int nnz = 0;
-	int *col_ptr = new int[n+1];
+	size_t nnz = 0;
+	size_t *col_ptr = new size_t [n+1];
 	feature_node *x_space;
 	prob_col->l = l;
 	prob_col->n = n;
